adding register fragment
