+++ app/src/main/res/values/strings.xml
@@ -20,4 +20,6 @@
     <string name="invalid_phone_number">Invalid phone number</string>
     <string name="too_short_password">Password must have at least 6 characters.</string>
     <string name="passwords_do_not_match">Passwords do not match</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
